public/requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.3
tqdm>=4.60.0
markdown>=3.3.0
openai>=1.0.0
pinecone-client>=3.0.0
flask>=2.0.0 